Convert from int based system to using decimals for currency to resolve issues found.
update UI to round decimals to 2 DP

+++ LotteryApp/Models/GameResultModel.cs
@@ -6,7 +6,7 @@ namespace LotteryApp.Models
     {
         public List<Player> Players { get; set; }
 
-        public int HouseProfit { get; set; }
+        public decimal HouseProfit { get; set; }
 
         public string Results { get; set; }
     }
